requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
cohere==4.37
googlesearch-python==1.2.3
gunicorn==20.1.0
openai==0.27.0
python-dotenv==0.19.0